content/canvas/test/test_toDataURL.lowercase.html

<!DOCTYPE HTML>
<title>Canvas test: toDataURL.lowercase - bug 401795</title>
<!-- Testing: toDataURL type is case-sensitive -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

var data = canvas.toDataURL('ImAgE/PnG');
ok(data =~ /^data:image\/png[;,]/, "data =~ /^data:image\\/png[;,]/");

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
